Minor fix to cbcBayesCombinePosteriors.py: Since numpy-1.12 float indices are no longer implicitly converted to integers

See merge request lscsoft/lalsuite!509

+++ lalinference/python/cbcBayesCombinePosteriors.py
@@ -186,7 +186,7 @@ else:
     
     testSize = sizeList[minIndex] / fracWeight[minIndex]
 
-    weightNum = np.around(fracWeight * testSize)
+    weightNum = np.around(fracWeight * testSize).astype(int)
     sizeOut = sum(weightNum)
     samplesOut = np.empty(sizeOut, dtype=datatypes)
